WIP: Authorization as wrappers
