+++ src/Level3/Processor/Wrapper/Authorization/AclAuthorizationWrapper.php
@@ -2,30 +2,27 @@
 
 namespace Level3\Processor\Wrapper\Authorization;
 
-use Level3\Messages\Processors\RequestProcessor;
 use Level3\Messages\Request;
 use Level3\Resources\ConfigParser;
 use Level3\Resources\Exceptions\ConfigError;
-use Level3\Processor\Wrapper\Authentication\Credentials;
+use Level3\Processor\Wrapper\Authentication\AuthenticatedCredentials;
 
-class AclAuthorizationProcessor extends AbstractAuthorizationProcessor
+class AclAuthorizationWrapper extends AbstractAuthorizationWrapper
 {
     private $config;
 
-    public function __construct(
-        RequestProcessor $processor,
-        ConfigParser $configParser
-    ) {
-        parent::__construct($processor);
+    public function __construct(ConfigParser $configParser)
+    {
         $this->config = $configParser->getConfig();
         if (!isset($this->config['acl']) || !isset($this->config['acl']['routes'])) {
-            throw new ConfigError('acl -> routes config sections have to be defined');
+            throw new ConfigError('acl -> routes config sections must be defined');
         }
     }
 
     protected function hasAccess(Request $request, $methodName)
     {
         $credentials = $request->getCredentials();
+        if (!($credentials instanceof AuthenticatedCredentials)) return false;
 
         foreach ($this->config['acl']['routes'] as $routeConfig) {
             if ($this->matches($request, $routeConfig)) {
@@ -35,7 +32,7 @@ class AclAuthorizationProcessor extends AbstractAuthorizationProcessor
         return false;
     }
 
-    private function credentialsHasAccess(Credentials $credentials, $routeConfig, $methodName, $path)
+    private function credentialsHasAccess(AuthenticatedCredentials $credentials, $routeConfig, $methodName, $path)
     {
         if (!isset($routeConfig['policies'][$methodName])) {
             return $this->hasDefaultAccess($routeConfig);
@@ -49,7 +46,7 @@ class AclAuthorizationProcessor extends AbstractAuthorizationProcessor
         return false;
     }
 
-    private function matchingRequirementsMeet(Credentials $credentials, $routeConfig, $methodName, $path)
+    private function matchingRequirementsMeet(AuthenticatedCredentials $credentials, $routeConfig, $methodName, $path)
     {
         if (!isset($routeConfig['policies'][$methodName]['matchings'])) {
             return true;
